Add CLI + config option for number of thread workers
 - Setting more than one worker enforces by default the --create-directories
 options. Otherwise we would have an issue of file/directory dependencies
 between threads.

WIP on issue #24

+++ inc/cloudmig/options.h
@@ -43,7 +43,7 @@ struct cloudmig_options
 {
     int                 flags;
     int                 trace_flags;
-    int                 nb_threads;
+    long int            nb_threads;
     const char          *src_profile;
     const char          *dest_profile;
     const char          *status_profile;
